requirements.txt
streamlit==1.29.0 
opencv-python==4.8.1
easyocr==1.7.1 
numpy==1.26.2
google-generativeai==0.8.1